pyproject.toml
[tool.poetry]
name = "cuebookassignment"
version = "0.1.0"
description = ""
authors = ["sakthiRatnam <sakthiratnam050@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.0.5"
djangorestframework = "3.12.4"
blinker = "^1.4"
APScheduler = "3.7.0"
django-allauth = "0.44.0"
django-cors-headers = "3.7.0"
django-cors-middleware = "1.5.0"
django-filter = "2.4.0"
django-rest-auth = "0.9.5"
django-robots = "4.0"
django-url-filter = "0.3.15"
requests = "^2.26.0"
gunicorn = "20.1.0"
psycopg2 = "2.8.6"
razorpay = "1.2.0"
reportlab = "3.5.67"
requests-oauthlib = "1.3.0"
rollbar = "0.15.2"
SecretStorage = "2.3.1"
uvicorn = "0.14.0"
celery = "5.1.2"
django-celery-beat = "2.2.1"
django-rest-swagger = "2.2.0"
redis = "^3.5.3"
django-prometheus = "^2.1.0"
exponent-server-sdk = "2.0.0"
geographiclib = "1.50"
geopy = "2.1.0"
oauthlib = "3.1.0"
a = "^1.0"
aiohttp = "^3.7.4"
boto = "^2.49.0"
scrape = "^0.11.1"
boto3 = "^1.18.51"
django-storages = "^1.11.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
